+++ qcodes/process/server.py
@@ -1,257 +1,18 @@
+"""Common Server process and ServerManager architecture."""
+
 import multiprocessing as mp
-import sys
-from datetime import datetime
-import time
-from traceback import print_exc, format_exc
+from traceback import format_exc
 from uuid import uuid4
 import builtins
 import logging
-import signal
-
-from .helpers import in_notebook
-
-MP_ERR = 'context has already been set'
 
 QUERY_WRITE = 'WRITE'
 QUERY_ASK = 'ASK'
 RESPONSE_OK = 'OK'
 RESPONSE_ERROR = 'ERROR'
 
-
-def set_mp_method(method, force=False):
-    """
-    An idempotent wrapper for multiprocessing.set_start_method.
-
-    The most important use of this is to force Windows behavior
-    on a Mac or Linux: set_mp_method('spawn')
-    args are the same:
-
-    method: one of:
-        'fork' (default on unix/mac)
-        'spawn' (default, and only option, on windows)
-        'forkserver'
-    force: allow changing context? default False
-        in the original function, even calling the function again
-        with the *same* method raises an error, but here we only
-        raise the error if you *don't* force *and* the context changes
-    """
-    try:
-        mp.set_start_method(method, force=force)
-    except RuntimeError as err:
-        if err.args != (MP_ERR, ):
-            raise
-
-    mp_method = mp.get_start_method()
-    if mp_method != method:
-        raise RuntimeError(
-            'unexpected multiprocessing method '
-            '\'{}\' when trying to set \'{}\''.format(mp_method, method))
-
-
-class QcodesProcess(mp.Process):
-
-    """
-    Modified multiprocessing.Process specialized to Qcodes needs.
-
-    - Nicer repr
-    - Automatic streaming of stdout and stderr to our StreamQueue singleton
-      for reporting back to the main process
-    - Ignore interrupt signals so that commands in the main process can be
-      canceled without affecting server and background processes.
-    """
-
-    def __init__(self, *args, name='QcodesProcess', queue_streams=True,
-                 daemon=True, **kwargs):
-        """
-        Construct the QcodesProcess but like Process, does not start it.
-
-        name: string to include in repr, and in the StreamQueue
-            default 'QcodesProcess'
-        queue_streams: should we connect stdout and stderr to the StreamQueue?
-            default True
-        daemon: should this process be treated as daemonic, so it gets
-            terminated with the parent.
-            default True, overriding the base inheritance
-        any other args and kwargs are passed to multiprocessing.Process
-        """
-        # make sure the singleton StreamQueue exists
-        # prior to launching a new process
-        if queue_streams and in_notebook():
-            self.stream_queue = get_stream_queue()
-        else:
-            self.stream_queue = None
-        super().__init__(*args, name=name, daemon=daemon, **kwargs)
-
-    def run(self):
-        """Executed in the new process, and calls the target function."""
-        # ignore interrupt signals, as they come from `KeyboardInterrupt`
-        # which we want only to apply to the main process and not the
-        # server and background processes (which can be halted in different
-        # ways)
-        signal.signal(signal.SIGINT, signal.SIG_IGN)
-
-        if self.stream_queue:
-            self.stream_queue.connect(str(self.name))
-        try:
-            super().run()
-        except:
-            # if we let the system print the exception by itself, sometimes
-            # it disconnects the stream partway through printing.
-            print_exc()
-        finally:
-            if (self.stream_queue and
-                    self.stream_queue.initial_streams is not None):
-                self.stream_queue.disconnect()
-
-    def __repr__(self):
-        """Shorter and more helpful repr of our processes."""
-        cname = self.__class__.__name__
-        r = super().__repr__()
-        r = r.replace(cname + '(', '').replace(')>', '>')
-        return r.replace(', started daemon', '')
-
-
-def get_stream_queue():
-    """
-    Convenience function to get a singleton StreamQueue.
-
-    note that this must be called from the main process before starting any
-    subprocesses that will use it, otherwise the subprocess will create its
-    own StreamQueue that no other processes know about
-    """
-    if StreamQueue.instance is None:
-        StreamQueue.instance = StreamQueue()
-    return StreamQueue.instance
-
-
-class StreamQueue:
-
-    """
-    Manages redirection of child process output for the main process to view.
-
-    Do not instantiate this directly: use get_stream_queue so we only make one.
-    One StreamQueue should be created in the consumer process, and passed
-    to each child process. In the child, we call StreamQueue.connect with a
-    process name that will be unique and meaningful to the user. The consumer
-    then periodically calls StreamQueue.get() to read these messages.
-
-    inspired by http://stackoverflow.com/questions/23947281/
-    """
-
-    instance = None
-
-    def __init__(self, *args, **kwargs):
-        """Create a StreamQueue, passing all args & kwargs to Queue."""
-        self.queue = mp.Queue(*args, **kwargs)
-        self.last_read_ts = mp.Value('d', time.time())
-        self._last_stream = None
-        self._on_new_line = True
-        self.lock = mp.RLock()
-        self.initial_streams = None
-
-    def connect(self, process_name):
-        """
-        Connect a child process to the StreamQueue.
-
-        After this, stdout and stderr go to a queue rather than being
-        printed to a console.
-
-        process_name: a short string that will clearly identify this process
-            to the user.
-        """
-        if self.initial_streams is not None:
-            raise RuntimeError('StreamQueue is already connected')
-
-        self.initial_streams = (sys.stdout, sys.stderr)
-
-        sys.stdout = _SQWriter(self, process_name)
-        sys.stderr = _SQWriter(self, process_name + ' ERR')
-
-    def disconnect(self):
-        """Disconnect a child from the queues and revert stdout & stderr."""
-        if self.initial_streams is None:
-            raise RuntimeError('StreamQueue is not connected')
-        sys.stdout, sys.stderr = self.initial_streams
-        self.initial_streams = None
-
-    def get(self):
-        """Read new messages from the queue and format them for printing."""
-        out = ''
-        while not self.queue.empty():
-            timestr, stream_name, msg = self.queue.get()
-            line_head = '[{} {}] '.format(timestr, stream_name)
-
-            if self._on_new_line:
-                out += line_head
-            elif stream_name != self._last_stream:
-                out += '\n' + line_head
-
-            out += msg[:-1].replace('\n', '\n' + line_head) + msg[-1]
-
-            self._on_new_line = (msg[-1] == '\n')
-            self._last_stream = stream_name
-
-        self.last_read_ts.value = time.time()
-        return out
-
-    def __del__(self):
-        """Tear down the StreamQueue either on the main or a child process."""
-        try:
-            self.disconnect()
-        except:
-            pass
-
-        if hasattr(type(self), 'instance'):
-            # so nobody else tries to use this dismantled stream queue later
-            type(self).instance = None
-
-        if hasattr(self, 'queue'):
-            kill_queue(self.queue)
-            del self.queue
-        if hasattr(self, 'lock'):
-            del self.lock
-
-
-class _SQWriter:
-    MIN_READ_TIME = 3
-
-    def __init__(self, stream_queue, stream_name):
-        self.queue = stream_queue.queue
-        self.last_read_ts = stream_queue.last_read_ts
-        self.stream_name = stream_name
-
-    def write(self, msg):
-        try:
-            if msg:
-                msgtuple = (datetime.now().strftime('%H:%M:%S.%f')[:-3],
-                            self.stream_name, msg)
-                self.queue.put(msgtuple)
-
-                queue_age = time.time() - self.last_read_ts.value
-                if queue_age > self.MIN_READ_TIME and msg != '\n':
-                    # long time since the queue was read? maybe nobody is
-                    # watching it at all - send messages to the terminal too
-                    # but they'll still be in the queue if someone DOES look.
-                    termstr = '[{} {}] {}'.format(*msgtuple)
-                    # we always want a new line this way (so I don't use
-                    # end='' in the print) but we don't want an extra if the
-                    # caller already included a newline.
-                    if termstr[-1] == '\n':
-                        termstr = termstr[:-1]
-                    try:
-                        print(termstr, file=sys.__stdout__)
-                    except ValueError:  # pragma: no cover
-                        # ValueError: underlying buffer has been detached
-                        # this may just occur in testing on Windows, not sure.
-                        pass
-        except:
-            # don't want to get an infinite loop if there's something wrong
-            # with the queue - put the regular streams back before handling
-            sys.stdout, sys.stderr = sys.__stdout__, sys.__stderr__
-            raise
-
-    def flush(self):
-        pass
+from .qcodes_process import QcodesProcess
+from .helpers import kill_queue
 
 
 class ServerManager:
@@ -606,12 +367,3 @@ class BaseServer:
             if name.startswith('handle_') and callable(getattr(self, name)):
                 handlers.append(name[len('handle_'):])
         return handlers
-
-
-def kill_queue(queue):
-    """Tear down a multiprocessing.Queue to help garbage collection."""
-    try:
-        queue.close()
-        queue.join_thread()
-    except:
-        pass
