+++ qcodes/tests/test_data.py
@@ -7,7 +7,7 @@ from qcodes.data.data_array import DataArray
 from qcodes.data.manager import get_data_manager, NoData
 from qcodes.data.data_set import (load_data, new_data, DataMode, DataSet,
                                   TimestampLocation)
-from qcodes.utils.helpers import killprocesses
+from qcodes.process.helpers import kill_processes
 from qcodes import active_children
 
 from .data_mocks import (MockDataManager, MockFormatter, FullIO, EmptyIO,
@@ -236,7 +236,7 @@ class TestDataArray(TestCase):
 
 class TestLoadData(TestCase):
     def setUp(self):
-        killprocesses()
+        kill_processes()
 
     def test_no_live_data(self):
         # live data with no DataManager at all
@@ -292,7 +292,7 @@ class TestLoadData(TestCase):
 
 class TestNewData(TestCase):
     def setUp(self):
-        killprocesses()
+        kill_processes()
         self.original_lp = DataSet.location_provider
 
     def tearDown(self):
@@ -361,7 +361,7 @@ class TestTimestampLocation(TestCase):
 
 class TestDataSet(TestCase):
     def tearDown(self):
-        killprocesses()
+        kill_processes()
 
     def test_constructor_errors(self):
         # no location - only allowed with load_data
